+++ lib/src/default_index/mutable.rs
@@ -31,9 +31,10 @@ use smallvec::SmallVec;
 use tempfile::NamedTempFile;
 
 use super::composite::CompositeIndex;
+use super::entry::{IndexEntry, IndexPosition, SmallIndexPositionsVec};
 use super::readonly::{DefaultReadonlyIndex, ReadonlyIndexSegment};
 use super::store::IndexLoadError;
-use super::{IndexEntry, IndexPosition, IndexSegment, SmallIndexPositionsVec};
+use super::IndexSegment;
 use crate::backend::{ChangeId, CommitId, ObjectId};
 use crate::commit::Commit;
 use crate::file_util::persist_content_addressed_temp_file;
